Kernel K1: move anim, save, autosave into kernel/ (pure code motion)
First slice of the kernel extraction (docs/PLATFORM.md §9): kernel/ is a
plain TS-source directory — deliberately NOT an npm workspace (the moved
modules have zero npm dependencies, so a workspace buys nothing yet while
a lockfile move would break CI and every worktree symlink).

- kernel/src/doc.ts: KernelDoc — the envelope (docId, title), the only
  doc fields kernel modules may read.
- git mv anim.ts, save.ts, autosave.ts → kernel/src/; save+autosave
  retyped BentoDoc → KernelDoc (type-only, zero runtime delta).
- docContentKey stays slides-side (in the autosave facade): 'what counts
  as content' is a per-app model question.
- Facades at the old slides/src paths re-export the kernel modules, so
  zero consumer files change and window.bento is untouched by
  construction.
- vite server.fs.allow ['..'] (dev-only) so the dev server may serve
  ../kernel.
- CI: standalone kernel typecheck step.

Byte-analysis of the built shell vs v1.0.8: everything outside the
bento-rt runtime blob is byte-identical (including #bento-doc); the
runtime is length-identical with 6 identifier-rename-only regions from
docContentKey's declaration moving. Conformance gate + SEEDS=100 rig
pass; dev-server smoke: boots, edits, kernel autosave writes recovery
snapshots, zero console errors.
