+++ slides/vite.config.ts
@@ -13,6 +13,10 @@ export default defineConfig({
   // compare against the release manifest. Single source: package.json.
   define: { __APP_VERSION__: JSON.stringify(pkg.version) },
   plugins: [...(process.env.SINGLEFILE ? [viteSingleFile()] : [])],
+  // Dev-only: allow serving ../kernel (shared TS-source kernel) — without a
+  // workspace root, vite's fs allow-list stops at the app dir. Build output
+  // is unaffected (rollup has no such restriction).
+  server: { fs: { allow: ['..'] } },
   build: {
     // Keep asset inlining aggressive; a Bento file must have zero external requests.
     assetsInlineLimit: 100_000_000,
